pyproject.toml
[project]
name = "resource-fula-ordboken"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    { name = "Kristoffer Andersson", email = "kristoffer.andersson@gu.se" },
]
dependencies = [
    "karp-lex-core>=0.6.2",
    "json-arrays>=0.14.1",
    "typer>=0.12.3",
    "tqdm>=4.66.2",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = ["ruff>=0.4.1", "mypy>=1.10.0", "pytest>=8.1.1", "pytest-cov>=5.0.0"]
